+++ examples/basic_code_area.rb
@@ -1,0 +1,31 @@
+require 'glimmer-dsl-libui'
+
+class BasicCodeArea
+  include Glimmer::LibUI::Application
+  
+  before_body do
+    @code = <<~CODE
+      # Greets target with greeting
+      def greet(greeting: 'Hello', target: 'World')
+        
+        puts "\#{greeting}, \#{target}!"
+      end
+      
+      greet
+      greet(target: 'Robert')
+      greet(greeting: 'Aloha')
+      greet(greeting: 'Aloha', target: 'Nancy')
+      greet(greeting: 'Howdy', target: 'Doodle')
+    CODE
+  end
+  
+  body {
+    window('Basic Code Area', 400, 300) {
+      margined true
+      
+      code_area(language: 'ruby', code: @code)
+    }
+  }
+end
+
+BasicCodeArea.launch