+++ lib/glimmer/libui/control_proxy.rb
@@ -401,4 +401,4 @@ module Glimmer
   end
 end
 
-Dir[File.expand_path('./control_proxy/*.rb', __dir__)].each {|f| require f}
+Dir[File.expand_path("./#{File.basename(__FILE__, '.rb')}/*.rb", __dir__)].each {|f| require f}